Avoid accidental gpui transitive dependency in collab
* Make Fs depend on Text, not vise versa
